# docs/openapi/banking-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Banking System API",
    "version": "1.0.0",
    "description": "REST API for customer, account, ledger, and transfer operations."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/customers": {
      "post": {
        "operationId": "createCustomer",
        "summary": "Create a customer",
        "x-required-scope": "banking.admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/customers/{customerId}": {
      "parameters": [
        {
          "name": "customerId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getCustomer",
        "summary": "Retrieve a customer",
        "x-required-scope": "banking.read",
        "responses": {
          "200": {
            "description": "Customer found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth 2.0 access token. Required scope is declared by x-required-scope."
      }
    },
...
    "responses": {
...
      "ApiError": {
        "description": "Request or business-rule failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
      "CreateCustomerRequest": {
        "type": "object",
        "required": ["fullName", "emailAddress"],
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "emailAddress": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "maxLength": 320
          }
        }
      },
      "Customer": {
        "type": "object",
        "required": ["customerId", "fullName", "emailAddress"],
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "emailAddress": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": ["timestamp", "status", "code", "message", "fieldErrors"],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}